motion service tests passing
