+++ services/motion/builtin/move_on_globe_test.go
@@ -3,21 +3,27 @@ package builtin
 import (
 	"context"
 	"errors"
-	"math"
 	"strings"
 	"testing"
+	"fmt"
 	"time"
 
 	"github.com/golang/geo/r3"
-	"github.com/google/uuid"
 	geo "github.com/kellydunn/golang-geo"
 	"go.viam.com/test"
 
+	"go.viam.com/rdk/components/camera"
 	_ "go.viam.com/rdk/components/register"
 	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/components/movementsensor"
 	"go.viam.com/rdk/services/motion"
 	"go.viam.com/rdk/spatialmath"
+	viz "go.viam.com/rdk/vision"
+	"go.viam.com/rdk/components/base"
+	_ "go.viam.com/rdk/components/register"
+	"go.viam.com/rdk/pointcloud"
+	"go.viam.com/rdk/services/vision"
+	"go.viam.com/rdk/testutils/inject"
 )
 
 func TestMoveOnGlobe(t *testing.T) {
@@ -27,8 +33,6 @@ func TestMoveOnGlobe(t *testing.T) {
 	// Near antarctica 🐧
 	gpsPoint := geo.NewPoint(-70, 40)
 	dst := geo.NewPoint(gpsPoint.Lat(), gpsPoint.Lng()+7e-5)
-	expectedDst := r3.Vector{X: 2662.16, Y: 0, Z: 0} // Relative pose to the starting point of the base; facing north, Y = forwards
-	epsilonMM := 15.
 	// create motion config
 	extra := map[string]interface{}{
 		"motion_profile": "position_only",
@@ -37,7 +41,7 @@ func TestMoveOnGlobe(t *testing.T) {
 	}
 
 	t.Run("Changes to executions show up in PlanHistory", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
 		defer closeFunc(ctx)
 
 		req := motion.MoveOnGlobeReq{
@@ -78,180 +82,8 @@ func TestMoveOnGlobe(t *testing.T) {
 		test.That(t, ph3, test.ShouldResemble, ph2)
 	})
 
-	t.Run("is able to reach a nearby geo point with empty values", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Destination:        dst,
-			Extra:              extra,
-		}
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-	})
-
-	t.Run("is able to reach a nearby geo point with a requested NaN heading", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Heading:            math.NaN(),
-			Destination:        dst,
-			Extra:              extra,
-		}
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-	})
-
-	t.Run("is able to reach a nearby geo point with a requested positive heading", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Heading:            10000000,
-			Destination:        dst,
-			Extra:              extra,
-		}
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-	})
-
-	t.Run("is able to reach a nearby geo point with a requested negative heading", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Heading:            -10000000,
-			Destination:        dst,
-			Extra:              extra,
-		}
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-	})
-
-	t.Run("is able to reach a nearby geo point when the motion configuration is empty", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Heading:            90,
-			Destination:        dst,
-			MotionCfg:          &motion.MotionConfiguration{},
-			Extra:              extra,
-		}
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-	})
-
-	t.Run("is able to reach a nearby geo point when the motion configuration nil", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Heading:            90,
-			Destination:        dst,
-			Extra:              extra,
-		}
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-	})
-
-	t.Run("ensure success to a nearby geo point", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		motionCfg := &motion.MotionConfiguration{PositionPollingFreqHz: 4, ObstaclePollingFreqHz: 1}
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			MovementSensorName: moveSensorResource,
-			Destination:        dst,
-			Obstacles:          []*spatialmath.GeoGeometry{},
-			MotionCfg:          motionCfg,
-			Extra:              extra,
-		}
-		planExecutor, err := ms.(*builtIn).newMoveOnGlobeRequest(ctx, req, nil, 0)
-		test.That(t, err, test.ShouldBeNil)
-
-		mr, ok := planExecutor.(*moveRequest)
-		test.That(t, ok, test.ShouldBeTrue)
-
-		test.That(t, mr.planRequest.Goal.Pose().Point().X, test.ShouldAlmostEqual, expectedDst.X, epsilonMM)
-		test.That(t, mr.planRequest.Goal.Pose().Point().Y, test.ShouldAlmostEqual, expectedDst.Y, epsilonMM)
-
-		planResp, err := mr.Plan(ctx)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, len(planResp.Path()), test.ShouldBeGreaterThan, 2)
-
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-
-		timeoutCtx, timeoutFn := context.WithTimeout(ctx, time.Second*15)
-		defer timeoutFn()
-		err = motion.PollHistoryUntilSuccessOrError(timeoutCtx, ms, time.Millisecond*5, motion.PlanHistoryReq{
-			ComponentName: req.ComponentName,
-			ExecutionID:   executionID,
-			LastPlanOnly:  true,
-		})
-		test.That(t, err, test.ShouldBeNil)
-	})
-
-	t.Run("go around an obstacle", func(t *testing.T) {
-		localizer, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
-		defer closeFunc(ctx)
-		planDeviationMM := 100.
-		motionCfg := &motion.MotionConfiguration{PositionPollingFreqHz: 0.0000001, LinearMPerSec: 0.2, AngularDegsPerSec: 60}
-
-		boxPose := spatialmath.NewPoseFromPoint(r3.Vector{X: 50, Y: 0, Z: 0})
-		boxDims := r3.Vector{X: 5, Y: 50, Z: 10}
-		geometries, err := spatialmath.NewBox(boxPose, boxDims, "wall")
-		test.That(t, err, test.ShouldBeNil)
-		geoGeometry := spatialmath.NewGeoGeometry(gpsPoint, []spatialmath.Geometry{geometries})
-		startPose, err := localizer.CurrentPosition(ctx)
-		test.That(t, err, test.ShouldBeNil)
-
-		req := motion.MoveOnGlobeReq{
-			ComponentName:      baseResource,
-			Destination:        dst,
-			MovementSensorName: moveSensorResource,
-			Obstacles:          []*spatialmath.GeoGeometry{geoGeometry},
-			MotionCfg:          motionCfg,
-			Extra:              extra,
-		}
-
-		executionID, err := ms.MoveOnGlobe(ctx, req)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, executionID, test.ShouldNotResemble, uuid.Nil)
-
-		timeoutCtx, timeoutFn := context.WithTimeout(ctx, time.Second*25)
-		defer timeoutFn()
-		err = motion.PollHistoryUntilSuccessOrError(timeoutCtx, ms, time.Millisecond*5, motion.PlanHistoryReq{
-			ComponentName: req.ComponentName,
-			ExecutionID:   executionID,
-			LastPlanOnly:  true,
-		})
-		test.That(t, err, test.ShouldBeNil)
-
-		endPose, err := localizer.CurrentPosition(ctx)
-		test.That(t, err, test.ShouldBeNil)
-		movedPose := spatialmath.PoseBetween(startPose.Pose(), endPose.Pose())
-		test.That(t, movedPose.Point().X, test.ShouldAlmostEqual, expectedDst.X, planDeviationMM)
-		test.That(t, movedPose.Point().Y, test.ShouldAlmostEqual, expectedDst.Y, planDeviationMM)
-	})
-
 	t.Run("fail because of obstacle", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
 		defer closeFunc(ctx)
 
 		// Construct a set of obstacles that entirely enclose the goal point
@@ -289,7 +121,7 @@ func TestMoveOnGlobe(t *testing.T) {
 	})
 
 	t.Run("check offset constructed correctly", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, gpsPoint, nil, 5)
 		defer closeFunc(ctx)
 		movementSensorInBase, err := ms.GetPose(ctx, resource.NewName(movementsensor.API, "test-gps"), "test-base", nil, nil)
 		test.That(t, err, test.ShouldBeNil)
@@ -312,7 +144,7 @@ func TestBoundingRegionsConstraint(t *testing.T) {
 	// Note: spatialmath.GeoPointToPoint(dst, origin) produces r3.Vector{1111.92, 0, 0}
 
 	t.Run("starting in collision with bounding regions works", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
 		defer closeFunc(ctx)
 
 		box, err := spatialmath.NewBox(spatialmath.NewZeroPose(), r3.Vector{2224, 2224, 2}, "")
@@ -333,7 +165,7 @@ func TestBoundingRegionsConstraint(t *testing.T) {
 	})
 
 	t.Run("starting outside of bounding regions fails", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
 		defer closeFunc(ctx)
 
 		box, err := spatialmath.NewBox(spatialmath.NewZeroPose(), r3.Vector{2222, 2222, 2}, "")
@@ -356,7 +188,7 @@ func TestBoundingRegionsConstraint(t *testing.T) {
 	})
 
 	t.Run("implicit success with no bounding regions", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
 		defer closeFunc(ctx)
 
 		req := motion.MoveOnGlobeReq{
@@ -371,7 +203,7 @@ func TestBoundingRegionsConstraint(t *testing.T) {
 	})
 
 	t.Run("fail to plan outside of bounding regions", func(t *testing.T) {
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
 		defer closeFunc(ctx)
 
 		box, err := spatialmath.NewBox(spatialmath.NewZeroPose(), r3.Vector{500, 500, 2}, "")
@@ -395,7 +227,7 @@ func TestBoundingRegionsConstraint(t *testing.T) {
 	t.Run("list of bounding regions - success case", func(t *testing.T) {
 		// string together multiple bounding regions, such that the robot must
 		// actually account for them in order to create a valid path
-		_, ms, closeFunc := createMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(ctx, t, origin, nil, 5)
 		defer closeFunc(ctx)
 
 		box1, err := spatialmath.NewBox(spatialmath.NewZeroPose(), r3.Vector{500, 500, 2}, "")
@@ -420,4 +252,157 @@ func TestBoundingRegionsConstraint(t *testing.T) {
 		_, err = ms.MoveOnGlobe(ctx, req)
 		test.That(t, err, test.ShouldBeNil)
 	})
+}
+
+
+func TestObstacleReplanningGlobe(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
+	ctx, cFunc := context.WithCancel(ctx)
+	defer cFunc()
+
+	gpsOrigin := geo.NewPoint(0, 0)
+	dst := geo.NewPoint(gpsOrigin.Lat(), gpsOrigin.Lng()+1e-5)
+	epsilonMM := 150.
+
+	type testCase struct {
+		name            string
+		getPCfunc       func(ctx context.Context, cameraName string, extra map[string]interface{}) ([]*viz.Object, error)
+		expectedSuccess bool
+		expectedErr     string
+		extra           map[string]interface{}
+	}
+
+	obstacleDetectorSlice := []motion.ObstacleDetectorName{
+		{VisionServiceName: vision.Named("injectedVisionSvc"), CameraName: camera.Named("injectedCamera")},
+	}
+
+	cfg := &motion.MotionConfiguration{
+		PositionPollingFreqHz: 1,
+		ObstaclePollingFreqHz: 5,
+		PlanDeviationMM: epsilonMM,
+		ObstacleDetectors: obstacleDetectorSlice,
+		 LinearMPerSec: 0.5,
+		 AngularDegsPerSec: 60,
+	}
+
+	extra := map[string]interface{}{"max_replans": 10, "max_ik_solutions": 1, "smooth_iter": 1, "motion_profile": "position_only"}
+	extraNoReplan := map[string]interface{}{"max_replans": 0, "max_ik_solutions": 1, "smooth_iter": 1}
+
+	// We set a flag here per test case so that detections are not returned the first time each vision service is called
+	testCases := []testCase{
+		{
+			name: "ensure no replan from discovered obstacles",
+			getPCfunc: func(ctx context.Context, cameraName string, extra map[string]interface{}) ([]*viz.Object, error) {
+				caseName := "test-case-1"
+				obstaclePosition := spatialmath.NewPoseFromPoint(r3.Vector{X: -1000, Y: -1000, Z: 0})
+				box, err := spatialmath.NewBox(obstaclePosition, r3.Vector{X: 10, Y: 10, Z: 10}, caseName)
+				test.That(t, err, test.ShouldBeNil)
+
+				detection, err := viz.NewObjectWithLabel(pointcloud.New(), caseName+"-detection", box.ToProtobuf())
+				test.That(t, err, test.ShouldBeNil)
+
+				return []*viz.Object{detection}, nil
+			},
+			expectedSuccess: true,
+			extra:           extraNoReplan,
+		},
+		{
+			name: "ensure replan due to obstacle collision",
+			getPCfunc: func(ctx context.Context, cameraName string, extra map[string]interface{}) ([]*viz.Object, error) {
+				caseName := "test-case-2"
+				// The camera is parented to the base. Thus, this will always see an obstacle 300mm in front of where the base is.
+				// Note: for CreateMoveOnGlobeEnvironment, the camera is given an orientation such that it is pointing left, not
+				// forwards. Thus, an obstacle in front of the base will be seen as being in +X.
+				obstaclePosition := spatialmath.NewPoseFromPoint(r3.Vector{X: 300, Y: 0, Z: 0})
+				box, err := spatialmath.NewBox(obstaclePosition, r3.Vector{X: 20, Y: 20, Z: 10}, caseName)
+				test.That(t, err, test.ShouldBeNil)
+
+				detection, err := viz.NewObjectWithLabel(pointcloud.New(), caseName+"-detection", box.ToProtobuf())
+				test.That(t, err, test.ShouldBeNil)
+
+				return []*viz.Object{detection}, nil
+			},
+			expectedSuccess: false,
+			expectedErr:     fmt.Sprintf("exceeded maximum number of replans: %d: plan failed", 0),
+			extra:           extraNoReplan,
+		},
+		{
+			name: "ensure replan reaching goal",
+			getPCfunc: func(ctx context.Context, cameraName string, extra map[string]interface{}) ([]*viz.Object, error) {
+				caseName := "test-case-3"
+				// This base will always see an obstacle 800mm in front of it, triggering several replans.
+				// However, enough replans should eventually get it to its goal.
+				obstaclePosition := spatialmath.NewPoseFromPoint(r3.Vector{X: 900, Y: 0, Z: 0})
+				box, err := spatialmath.NewBox(obstaclePosition, r3.Vector{X: 1, Y: 1, Z: 10}, caseName)
+				test.That(t, err, test.ShouldBeNil)
+
+				detection, err := viz.NewObjectWithLabel(pointcloud.New(), caseName+"-detection", box.ToProtobuf())
+				test.That(t, err, test.ShouldBeNil)
+
+				return []*viz.Object{detection}, nil
+			},
+			expectedSuccess: true,
+			extra:           extra,
+		},
+	}
+
+	testFn := func(t *testing.T, tc testCase) {
+		t.Helper()
+
+		calledPC := false
+		pcFunc := func(ctx context.Context, cameraName string, extra map[string]interface{}) ([]*viz.Object, error) {
+			if !calledPC {
+				calledPC = true
+				return []*viz.Object{}, nil
+			}
+			return tc.getPCfunc(ctx, cameraName, extra)
+		}
+
+		_, ms, closeFunc := CreateMoveOnGlobeEnvironment(
+			ctx,
+			t,
+			gpsOrigin,
+			nil,
+			5000,
+		)
+		defer closeFunc(ctx)
+
+		srvc, ok := ms.(*builtIn).visionServices[cfg.ObstacleDetectors[0].VisionServiceName].(*inject.VisionService)
+		test.That(t, ok, test.ShouldBeTrue)
+		srvc.GetObjectPointCloudsFunc = pcFunc
+
+		req := motion.MoveOnGlobeReq{
+			ComponentName:      resource.NewName(base.API, baseName),
+			Destination:        dst,
+			MovementSensorName: resource.NewName(movementsensor.API, moveSensorName),
+			MotionCfg:          cfg,
+			Extra:              tc.extra,
+		}
+		executionID, err := ms.MoveOnGlobe(ctx, req)
+		test.That(t, err, test.ShouldBeNil)
+
+		timeoutCtx, timeoutFn := context.WithTimeout(ctx, time.Minute*5)
+		defer timeoutFn()
+		err = motion.PollHistoryUntilSuccessOrError(timeoutCtx, ms, time.Millisecond*5, motion.PlanHistoryReq{
+			ComponentName: req.ComponentName,
+			ExecutionID:   executionID,
+			LastPlanOnly:  true,
+		})
+
+		if tc.expectedSuccess {
+			test.That(t, err, test.ShouldBeNil)
+		} else {
+			test.That(t, err, test.ShouldNotBeNil)
+			test.That(t, err.Error(), test.ShouldEqual, tc.expectedErr)
+		}
+	}
+
+	for _, tc := range testCases {
+		c := tc // needed to workaround loop variable not being captured by func literals
+		t.Run(c.name, func(t *testing.T) {
+			// These cannot be parallel or the `defer`s will cause failures
+			testFn(t, c)
+		})
+	}
 }
